+++ src/core/WfcRhoTools.cpp
@@ -3,7 +3,7 @@
 #include <stdexcept>
 
 namespace WfcToRho {
-	void BoundFermiGas::calcWeights(int nElec, const double* energies, double* weights, NormalizationScheme norm) {
+	void BoundFermiGas::calcWeights(size_t nElec, const double* energies, double* weights, NormalizationScheme norm) {
 		if(norm == NormalizationScheme::UNNORMALIZED)
 			throw std::runtime_error("WfcToRho::BoundFermiGas Cannot use UNNORMALIZED scheme with BoundFermiGas");
 
@@ -11,19 +11,19 @@ namespace WfcToRho {
 		double maxE = vtls::max(nElec, energies);
 		//Find center of Fermi slab, and set new 0-energy accoridngly, and convert to Fermi energy difference
 		double bottom = (maxE + minE) / 2 - ef / 2;
-		for (int i = 0; i < nElec; i++)
+		for (size_t i = 0; i < nElec; i++)
 			weights[i] = ef - (energies[i] - bottom);
 		//Get factored prefactor
 		double fact = 0;
-		for (int i = 0; i < nElec; i++)
+		for (size_t i = 0; i < nElec; i++)
 			fact += weights[i];
 		fact = 2.0 / (3.0 * PhysCon::pi) * PhysCon::me * ef / (PhysCon::hbar * PhysCon::hbar) * nElec / fact;
 		//Combine factored prefactor with energy difference
-		for (int i = 0; i < nElec; i++)
+		for (size_t i = 0; i < nElec; i++)
 			weights[i] *= fact;
 	}
 
-	void SemiInfiniteFermiGas::calcWeights(int nElec, const double* energies, double* weights, NormalizationScheme norm) {
+	void SemiInfiniteFermiGas::calcWeights(size_t nElec, const double* energies, double* weights, NormalizationScheme norm) {
 		if(norm == NormalizationScheme::NORMALIZED)
 			throw std::runtime_error("WfcToRho::SemiInfiniteFermiGas expects a wavefunction which has a normalization dicated by boundary conditions, not the total norm.");
 
@@ -34,7 +34,7 @@ namespace WfcToRho {
 		// get energy range boundaries relative to bottom
 		double* energyRangeBoundaries = (double*) sq_malloc(sizeof(double) * (nElec+1));
 		energyRangeBoundaries[0] = 0.0;
-		for (int i = 0; i < nElec-1; i++){
+		for (size_t i = 0; i < nElec-1; i++){
 			if(energies[i] > energies[i+1])
 				throw std::runtime_error("WfcToRho::SemiInfiniteFermiGas::calcWeights: energies not sorted");
 			energyRangeBoundaries[i + 1] = (energies[i] + energies[i+1])/2.0 - bottom;
@@ -46,7 +46,7 @@ namespace WfcToRho {
 		// density between energies \propto (Ef-E1)^2-(Ef-E0)^2
 		// each state has density = N * [(Ef-E1)^2-(Ef-E0)^2] / Ef^2
 		double n0 = std::sqrt(2)/(3.0*PhysCon::pi*PhysCon::pi)*std::pow(ef * PhysCon::me/PhysCon::hbar/PhysCon::hbar, 1.5);
-		for(int i = 0; i < nElec; i++)
+		for(size_t i = 0; i < nElec; i++)
 			weights[i] = n0 * (std::pow(ef - energyRangeBoundaries[i], 2) - std::pow(ef - energyRangeBoundaries[i+1], 2)) / ef / ef;
 
 		sq_free(energyRangeBoundaries);
@@ -76,7 +76,7 @@ namespace WfcToRho {
 		sq_free(fE);
 	}
 
-	void FromDOS::calcWeights(int nElec, const double* energies, double* weights, NormalizationScheme norm) {
+	void FromDOS::calcWeights(size_t nElec, const double* energies, double* weights, NormalizationScheme norm) {
 		//approximate effective width of well... will need to be reconsidered if using non square-ish wells
 		//corrects for lost normalized density for larger wells (densities should be O(1))
 		//double leff = PhysCon::hbar * 2.0 * PhysCon::pi * nElec / (2.0 * std::sqrt(2.0 * ef * PhysCon::me));
@@ -91,17 +91,17 @@ namespace WfcToRho {
 		}
 
 		//sort energy
-		int* idx = (int*) sq_malloc(sizeof(int)*nElec);
+		size_t* idx = (size_t*) sq_malloc(sizeof(size_t)*nElec);
 		double* sortedEnergies = (double*) sq_malloc(sizeof(double) * nElec);
 		vtls::copyArray( nElec, energies, sortedEnergies);
 
-		for (int i = 0; i < nElec; i++)
+		for (size_t i = 0; i < nElec; i++)
 			idx[i] = i;
 		vtls::insertSort_idxs(nElec, sortedEnergies, idx);
 
 		//represented energies are half-way between adjacent energies
 		//eg, if we have energy states E = 0, 1, 3, then the state of energy 1 represents energies 2 <- 0.5
-		for (int i = 1; i < nElec - 1; i++) {
+		for (size_t i = 1; i < nElec - 1; i++) {
 			weights[idx[i]] = leff * (dosISpline((sortedEnergies[i] + sortedEnergies[i + 1]) / 2.0) - dosISpline((sortedEnergies[i] + sortedEnergies[i - 1]) / 2.0));
 		}
 
@@ -116,17 +116,17 @@ namespace WfcToRho {
 		sq_free(sortedEnergies);
 	}
 
-	void Density::calcRawRho(int nPts, int nElec, const double* weights, const std::complex<double>* psi, double* psi2_work, double* rho){
+	void Density::calcRawRho(size_t nPts, size_t nElec, const double* weights, const std::complex<double>* psi, double* psi2_work, double* rho){
 		std::fill_n(rho, nPts, 0);
 		vtls::normSqr(nPts * nElec, psi, psi2_work);
-		for (int i = 0; i < nElec; i++) {
+		for (size_t i = 0; i < nElec; i++) {
 			double pref = weights[i];
-			for (int j = 0; j < nPts; j++)
+			for (size_t j = 0; j < nPts; j++)
 				rho[j] += psi2_work[i * nPts + j] * pref;
 		}
 	}
 
-	void DirectDensity::calcRho(int nPts, int nElec, double dx, double* rho) {
+	void DirectDensity::calcRho(size_t nPts, size_t nElec, double dx, double* rho) {
 		return;
 	}
 
@@ -137,7 +137,7 @@ namespace WfcToRho {
 			sq_free(thinning);
 	}
 
-	void CylindricalDensity::doFirst(int nPts, double dx) {
+	void CylindricalDensity::doFirst(size_t nPts, double dx) {
 		first = false;
 		if(thinning)
 			sq_free(thinning);
@@ -146,21 +146,21 @@ namespace WfcToRho {
 		std::fill_n(thinning, nPts, 1.0);
 
 		if(radius > dx/2){
-			startIndex = std::max(0, (int)std::floor((center + radius - minX) / dx));
+			startIndex = std::max((size_t)0, (size_t)std::floor((center + radius - minX) / dx));
 			endIndex = nPts;
 		}
 		else if(radius < -dx/2){
 			startIndex = 0;
-			endIndex = std::min(nPts, (int)std::ceil((center + radius - minX) / dx));
+			endIndex = std::min(nPts, (size_t)std::ceil((center + radius - minX) / dx));
 		}
 		else
 			throw std::runtime_error("CylindricalDensity: Radius must be larger than half of grid size.");
 		
-		for(int i = startIndex; i < endIndex; i++)
+		for(size_t i = startIndex; i < endIndex; i++)
 			thinning[i] = radius / (i * dx - center - minX);
 	}
 
-	void CylindricalDensity::calcRho(int nPts, int nElec, double dx, double* rho) {
+	void CylindricalDensity::calcRho(size_t nPts, size_t nElec, double dx, double* rho) {
 		if (first)
 			doFirst(nPts, dx);
 
@@ -174,7 +174,7 @@ namespace WfcToRho {
 			delete conv;
 	}
 
-	void GaussianSmoothedDensity::calcRho(int nPts, int nElec, double dx, double* rho) {
+	void GaussianSmoothedDensity::calcRho(size_t nPts, size_t nElec, double dx, double* rho) {
 		if (first) {
 			//Initialize variables
 			if(tempRho)
@@ -184,7 +184,7 @@ namespace WfcToRho {
 			first = false;
 
 			//Initialize Gaussian mask (in k space)
-			for (int i = 0; i < nPts / 2; i++) {
+			for (size_t i = 0; i < nPts / 2; i++) {
 				mask[i] = 1.0 / (sig/dx * std::sqrt(2.0 * PhysCon::pi)) * std::exp(-0.5 / (sig * sig) * (i * i * dx * dx));
 				mask[nPts - i - 1] = 1.0 / (sig/dx * std::sqrt(2.0 * PhysCon::pi)) * std::exp(-0.5 / (sig * sig) * ((i+1) * (i+1) * dx * dx));
 			}
